WIP - Explore streaming library
This branch explores implementing a reactive streams inspired streaming
API ontop of Pantomime.

The main design here is a directed acyclic graph, at least in terms of
the datastructures given Rust's ownership system. Ownership transfers up
and down this graph depending upon the event that occurred.

Asynchronous boundaries will be added soon that are implemented using
actors. This can be done because you can clone an ActorRef and send it
upstream and downstream simultaneously, allowing independent progress to
be made.

Note the DAG part. Cycles can still be implemented but are intended to
be modeled more directly, as it A) avoids having a runtime graph
interpreter, B) can be better expressed using Rust and explicitly wiring up actors.