Be more prescriptive in application in application structure
Entry into Pantomime is now a single `ActorSystem::spawn` method that
takes an actor.

This actor, conventionally called a reaper, is then responsible for
responding to signals, starting and stopping actors, etc.

This unifies the API -- all spawning that a user's application can do is
done within the context of an actor.

This also provides a nice API for testing actors. In a test, define a
reaper that does some assertions, and then stops itself. If the reaper
failed, an `io::Error` will be returned, otherwise `io::Ok` is returned.

Not all of the tests have been updated yet, but some have. Also see
ping-pong for an example application using this new style.