Simplify the implementation.
Given that the required feature set is mostly figured out, this now
refactors the types to be more straight forward.

This also does away with the sharding concept for now. This does mean an
increase in memory usage for large amounts of actors (10M or more), as
mailboxes are no longer shared. I'm unsure of a solution to this
problem at the moment, but I think that the clarity in design is worth
it.