Provide a simple configuration mechanism
This provides an environment-variable based configuration mechanism that
leverages FromStr/parse.

The intent is for libraries e.g. Pantomime to be specify defaults in
code, applications to specify defaults in main (including overriding any
library defaults), and for the operator to be able to override anything
at runtime via the environment.