Overhaul panic handling
Don't rely on catch_unwind. Instead, we use deferred to check for thread
panicking and message ourselves a Stop(fail = true) if that's the case.

In general, note that panics should be avoided, but sometimes unforseen
things happen hence it's necessary to consider them in Pantomime.

Supervisor and Streams still need to be extended to use this mechanism,
that will happen in a separate commit.