Use associated types where appropriate
For stream logic, an associated type is used for the control message, as
it's implementation specific. However, the input and output types are
call-site specific, so they remain as type parameters.

Similarly for actors, since the message type is implementation specific,
it now uses an associated type instead of a type parameter.