stream: Implement SourceQueue
This commit adds an initial SourceQueue implementation. This allows
streams to be spawned, and separately elsewhere for items to be pushed
into the stream.

This initial commit is a rough cut -- it doesn't properly clean up, and
isn't ideal from a performance perspective. However, it's interface is
defined well enough to allow testing in other applications to collect
feedback.