streams: Implement MapConcat
This adds `Flow::map_concat` and `Stream::map_concat` which takes a
function that consumes an `A` and produces an iterator of `B`s.

This `B`s are emitted serially until the iterator is exhausted, at which
point upstream is pulled for another element.

This can be used to e.g. turn a `Source<Vec<A>>` into a `Source<A>`.